refactor: split Eq/Ord and From/TryFrom impls out of lib.rs into comparisons.rs and conversions.rs
